fix(util): move propTypes to util folder
