+++ src/util/fontAwesomeIconPropTypes.js
